Allow re-transmission with same FCnt value.
This might happen when an ACK has been send to the node,
but the node did not receive it. In this case the node will
re-transmit the same packet (and thus with the same FCnt).
In this case the FCnt has not been incremented, but the packet
needs to be ACK-ed by the server (again).