Implement GetNodeSessionForPHYPayload.
One big drawback is that we can’t notify the application about incorrect MIC or FCnt as these values are part of the network-session lookup. In other words: in order to know to which network-session a DevAddr belongs, the FCnt and MIC must be correct. The advantage is that the available address space has increased drastically.