Fix NbTrans in case of early packetloss.
E.g. if there would be packetloss of one packet of the first three, this would be seen as 33% loss, which would increase NbTrans rapidly.

In the new solution the packetloss function returns 0 until the history table has been filled (20 packets). Also the ADR engine will lower the NbTrans even if there is no data-rate or txpower change (this was handled incorrectly before).