Fixed the "index out of range" error when scanning for conflicting MAC commands.
In setMACCommands, the code checks to see whether the MAC payload contains any conflicting MAC commands and if so, removes the conflicting ones. However, the code was modifying the array while ascending through it in the for loop, which can cause bugs as the array range calculated at the beginning of the for loop is now wrong. This fix scans the array in descending order, thus avoiding any problems.

Fixes #323