Add gateway specific context field + refactor downlink timing.
This makes the scheduling of downlinks more abstract so that it becomes
easier to support different packet-forwarders. The gateway internal
timestamp will move to the context object (binary blob) and to schedule
a Class-A downlink, the NS only needs to return the same uplink context
blob + a delay instruction with the delay relative to the uplink
context.

This is implemented in a backwards compatible way, but the old fields
will be removed in the next major version (v3).