Remove MaxSupportedDR field from device-session.
This field used to be there to support devices that did not
implement add specified DRs. The reason to remove this is
that devices must support all available DRs, but also that the
max supported DR can be specified in the service-profile.

By removing this from the device-session, the max allowed DR from
the service-profile is always used, meaning it can be changed any
time, without the need to re-activate the device.