Implement Azure Service-Bus Queue / IoT Hub backend.
This is based on the work of #386 by @asanchezdelc with the modification
that this backend doesn't subscribe directly to the IoT Hub interface,
but uses a Service-Bus Queue so that events gets distributed across
multiple LoRa Server instances.

TODO: evaluate using the Cloud-to-Device REST API endpoint instead of
external library. This is currently blocked by incomplete API
documentation.

Closes #386.