Extend meta-data sent to NC for uplink + add downlink meta-data.
The downlink meta-data is sent when the gateway acknowledges the downlink
scheduling (the NC might use the meta-data for billing purposes and we
don't want to charge for downlinks that are not actually emitted by the
gateway).
Therefore the saved downlink-frames has been modified to store the last
downlink + alternative downlinks. In case of a gateway ACK, this means
we still have the context of the ACKed downlink (tx-parameters +
downlink payload).