Remove api package. Update scripts to use chirpstack-api Protobuf files.
The API files have been moved to
https://github.com/brocaar/chirpstack-api. As some internal Protobuf
files were still depending on the files within the api directory, this
was never cleaned up. This will remove the last "api" traces and update
the Makefile to pull in the chirpstack-api Protobuf (matching the Go
module version) and update the Protobuf include paths.