Cleanup gateway stats code.
Stats were previously stored in SQL, then migrated to Redis and in v3
migrated to the AS. This removes the storage functions (except
retrieving the stats from Redis) as they are no longer being used. It
also cleans up the database table.

The stats retrieval function remains so that the stats migration to the
AS is still possible, when upgrading from an older v3 version.