Log the actual downlink frame accepted by gw.
Previously, the first attempt would be logged. However, there is a
possibility that the first attempt fails (e.g. collision with other
items in the gateway queue), and that the downlink is scheduled using
RX2 parameters (potentially with more / less mac-commands as a different
data-rate might have different payload size constraints).

With this change, a downlink is logged only when the gateway sends a
positive acknowledgement.