Refactor sending downlink + tx ack.
Instead of sending one frame and storing the retry possibilities, this
sends all downlink possibilities at once and expects an array of ack
items. In case of a retry, this saves one roundtrip reducing the risk of
a failed downlink due to latency.

To remain backwards compatible, the default is to set the old and new
fields. This has a slight overhead in bandwith usage, but makes it
possible to migrate the Chirpstack Gateway Bridge to a newer version.
This behavior can be controlled through the `multi_downlink_feature`
setting.