Cleanup Geolocation Server integration.
Geolocation is now provided through a LoRa Cloud integration in
ChirpStack Application Server. Alternatively, all the meta-data
(fine-timestamp, SNR, RSSI) needed for geolocation is already exposed
by ChirpStack Application Server so that geolocation through an external
service is still a possibility.

The advantage of performing geolocation in the Application Server is
that it can use a combination of TDOA, RSSI, Wifi, GNSS and potentially
other sources for geolocation. For geolocation using Wifi, GNSS, ... the
payload must be decrypted first, which can only be done by the
Application Server.