Enable config option for default roaming server (configuration).
This makes it possible to 'auto discover' roaming agreements using DNS
and mutual trusted TLS certificates signed with a common CA certificate.

On OTAA:

* the fNS will request the hNS NetID through a join-server
* discover the hNS using DNS (using resolve_netid_domain_suffix)
* start a passive-roaming session using PRStartReq
* store the passive-roaming session for the activated DevAddr

Note: passive-roaming using the default server configuration only works
when the passive-roaming session is started with an OTAA join-request.
Otherwise, the fNS has no knowledge about the hNS NetID because it can't
be derrived when only the DevAddr is given.