Link gw to service-profile + add gws_private option.
The purpose of linking a gateway with a service-profile is such that we
can set additional settings such as if gateways under a given
service-profile are private to the given service-profile (and thus only
can be used by devices under the same service-profile).

This also implements the filtering of uplink frames based on the above
gws_private field. When the gws_private is set to true, received uplinks
will only be available to devices under the same service-profile.
Assigning a service-profile to a gateway is currently optional and when
not assigned, the gateway is considered public.